cross platform: replace long to int32

+++ lib/Backend/arm64/EncoderMD.cpp
@@ -6,7 +6,7 @@
 #include "ARMEncode.h"
 
 bool
-EncoderMD::EncodeImmediate16(long constant, DWORD * result)
+EncoderMD::EncodeImmediate16(int32 constant, DWORD * result)
 {
     if (constant > 0xFFFF)
     {
